Add a safe Drop impl for Texture
Texture is now a wrapper around Option<SdlTexture> because
SdlTexture::destroy() takes ownership of self. This is annoying because
it increases the size of a Texture, and we must unwrap() everywhere, but
hopefully I can fix at least the former by submitting a PR to change the
pointers held in sdl2 structs to std::ptr::NonNull<T>. The latter could
also be fixed when unwrap_unchecked is stabilised, but I'm not sure I'm
comfortable with that much unsafe code.